portfolio/requirements.txt
asgiref==3.2.7
boto3==1.13.7
botocore==1.16.7
Django==3.0.6
django-crispy-forms==1.9.1
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
Pillow==7.1.2
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
